Reorginize tests in file manual-and-auto-linking-when-creating-new-marcBib.cy.js
